reorganize and remove peer on send failure
